2003-01-20  Al Riddoch  <alriddoch@zepler.org>
	* src/Segment.cpp, src/Segment.h, src/Terrain.cpp, src/Terrain.h:
	  Switch to using floats instead of doubles internally for performance.

+++ Mercator/Segment.cpp
@@ -8,18 +8,18 @@
 
 namespace Mercator {
 
-Segment::Segment(int res) : m_res(res), m_points(new double[res * res])
+Segment::Segment(int res) : m_res(res), m_points(new float[res * res])
 {
 }
 
-void Segment::populate(double nn, double fn, double ff, double nf)
+void Segment::populate(float nn, float fn, float ff, float nf)
 {
     // A somewhat wacky proof of concept interpolation
     for(int i = 0; i < m_res; ++i) {
-        double near = nn + ((double)i / m_res) * (fn - nn);
-        double far = nf + ((double)i / m_res) * (ff - nf);
+        float near = nn + ((float)i / m_res) * (fn - nn);
+        float far = nf + ((float)i / m_res) * (ff - nf);
         for(int j = 0; j < m_res; ++j) {
-            m_points[j * m_res + i] = near + ((double)j / m_res) * (far - near);
+            m_points[j * m_res + i] = near + ((float)j / m_res) * (far - near);
         }
     }
 }
